pyproject.toml
[project]
name = "web-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.44",
    "fastapi>=0.118.0",
    "gunicorn>=21.2.0",
    "load-dotenv>=0.1.0",
    "loguru>=0.7.3",
    "paramiko>=4.0.0",
    "psycopg2-binary>=2.9.7",
    "pydantic[email]>=2.11.9",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "supabase>=2.20.0",
    "tencentcloud-sdk-python-common>=3.0.1469",
    "tencentcloud-sdk-python-dnspod>=3.0.1459",
    "uv>=0.8.22",
    "uvicorn>=0.37.0",
]
